+++ src/app/Models/BaseModel.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Different\Dwfw\app\Models;
+
+use Different\Dwfw\app\Models\Traits\BaseDwfwTrait;
+use Illuminate\Database\Eloquent\Model;
+
+class BaseModel extends Model
+{
+    use BaseDwfwTrait;
+}