+++ src/config/checkIp.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+
+    'allow_list' => ['*'],
+    'block_list' => [],
+];