+++ src/config/dwfw.php
@@ -1,0 +1,3 @@
+<?php
+
+return ['version' => '0.10.14'];