+++ src/resources/lang/backpack/hu/base.php
@@ -1,0 +1,70 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Backpack\Base Language Lines
+    |--------------------------------------------------------------------------
+    */
+
+    'registration_closed'    => 'Regisztráció nem lehetséges.',
+    'no_email_column'        => 'A felhasználóhoz nem tartozik e-mail cím.',
+    'first_page_you_see'     => 'Kezdő oldal bejelentkezés után',
+    'login_status'           => 'Bejelentkezési státusz',
+    'logged_in'              => 'Bejelentkezve!',
+    'toggle_navigation'      => 'Navigáció kapcsoló',
+    'administration'         => 'ADMINISZTRÁCIÓ',
+    'user'                   => 'FELHASZNÁLÓ',
+    'logout'                 => 'Kijelentkezés',
+    'login'                  => 'Bejelentkezés',
+    'register'               => 'Regisztráció',
+    'name'                   => 'Név',
+    'email_address'          => 'E-mail cím',
+    'password'               => 'Jelszó',
+    'old_password'           => 'Régi jelszó',
+    'new_password'           => 'Új jelszó',
+    'confirm_password'       => 'Jelszó megerősítése',
+    'remember_me'            => 'Emlékezz rám',
+    'forgot_your_password'   => 'Elfelejtett jelszó?',
+    'reset_password'         => 'Jelszó helyreállítása',
+    'send_reset_link'        => 'Jelszó helyreállítási link küldése',
+    'click_here_to_reset'    => 'A linkre kattintva a jelszó helyreállítható',
+    'change_password'        => 'Jelszó megváltoztatása',
+    'unauthorized'           => 'Nem megfelelő jogosultság.',
+    'dashboard'              => 'Dashboard',
+    'handcrafted_by'         => 'Készítette',
+    'powered_by'             => 'Felhasználva',
+    'my_account'             => 'Profilom',
+    'update_account_info'    => 'Profil adatainak frissítése',
+    'save'                   => 'Mentés',
+    'cancel'                 => 'Mégse',
+    'error'                  => 'Hiba',
+    'success'                => 'Siker',
+    'warning'                => 'Figyelmeztetés',
+    'notice'                 => 'Megjegyzés',
+    'old_password_incorrect' => 'A régi jelszó nem megfelelő.',
+    'password_dont_match'    => 'A megadott jelszavak nem egyeznek.',
+    'password_empty'         => 'Mindkét jelszó mező kitöltése kötelező.',
+    'password_updated'       => 'Jelszó frissítve.',
+    'account_updated'        => 'Profil sikeresen frissítve.',
+    'unknown_error'          => 'Ismeretlen hiba. Kérjük próbálja újra.',
+    'error_saving'           => 'Hiba mentés közben. Kérjük próbálja újra.',
+    'welcome'                => 'Üdvözöljük!',
+    'use_sidebar'            => 'Az oldalsáv használatával hozhat létre, szerkeszthet vagy törölhet tartalmakat.',
+
+    'password_reset' => [
+        'greeting' => 'Hello!',
+        'subject'  => 'Jelszó-visszaállítási kérelem',
+        'line_1'   => 'Az ehhez az e-mail címhez tartozó felhasználóhoz jelszó-helyreállítási kérelem érkezett.',
+        'line_2'   => 'A jelszó visszaállítás az alábbi gombra kattintva lehetséges:',
+        'button'   => 'Jelszó visszaállítása',
+        'notice'   => 'Amennyiben nem ön igényelte a jelszó visszaállítását, nincs további teendő.',
+    ],
+
+    'step'                 => 'Lépés',
+    'confirm_email'        => 'Email cím megerősítése',
+    'choose_new_password'  => 'Új jelszó választása',
+    'confirm_new_password' => 'Új jelszó megerősítése',
+
+];