+++ src/resources/lang/backpack/hu/permissionmanager.php
@@ -1,0 +1,30 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Permission Manager Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used for Laravel Backpack - Permission Manager
+    | Author: Lúdio Oliveira <ludio.ao@gmail.com>
+    |
+    */
+    'name'                  => 'Név',
+    'role'                  => 'Szerep',
+    'roles'                 => 'Szerepek',
+    'roles_have_permission' => 'Szerepek a megfelelő jogosultsággal',
+    'permission_singular'   => 'jogosultság',
+    'permission_plural'     => 'jogosultságok',
+    'user_singular'         => 'Felhasználó',
+    'user_plural'           => 'Felhasználók',
+    'email'                 => 'Email',
+    'extra_permissions'     => 'Extra Jogosultságok',
+    'password'              => 'Jelszó',
+    'password_confirmation' => 'Jelszó megerősítése',
+    'user_role_permission'  => 'Felhasználói szerep jogosultságai',
+    'user'                  => 'Felhasználó',
+    'users'                 => 'Felhasználók',
+    'guard_type'            => 'Guard Típusa',
+
+];