+++ src/resources/lang/backpack/hu/settings.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Settings Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used for Laravel Backpack - Settings
+    |
+    */
+    'name'             => 'Név',
+    'value'            => 'Érték',
+    'description'      => 'Leírás',
+    'setting_singular' => 'Beállítás',
+    'setting_plural'   => 'Beállítások',
+
+];