+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.10.14] - 2020-09-02
+### added
+- Filters for Log crud
+- Hungarian translations for backpackforlaravel package
+- BaseModel
+- BaseDwfwApiFormRequest
+- Disable debugbar middleware
+- Ip restriction middleware
+- Upgrade command (php artisan dwfw:upgrade)
+
+### changed
+- Breaking change: New abstract method in BaseCrudController
+
 ## [0.10.13] - 2020-08-27
 ### fixed
 - Dwfwuser Trait file route added missing parameter